lib/baloon.c: fix a compiler warning
A warning emitted by gcc-4.1.2 on x86:

balloon.c: In function ‘ploop_balloon_change_size’:
balloon.c:340: warning: format ‘%ld’ expects type ‘long int’, but argument 3 has type ‘off_t’
balloon.c:340: warning: format ‘%ld’ expects type ‘long int’, but argument 4 has type ‘off_t’

+++ lib/balloon.c
@@ -337,7 +337,7 @@ int ploop_balloon_change_size(const char *device, int balloonfd, off_t new_size)
 	new_size = ((new_size << 9) + st.st_blksize - 1) & ~(st.st_blksize - 1);
 
 	ploop_log(0, "change balloon size old_size=%ld new_size=%ld",
-			old_size, new_size);
+			(long)old_size, (long)new_size);
 
 	pfiemap = fiemap_alloc(128);
 	freemap = freemap_alloc(128);
